moved code away from the generated file
