+++ go/vt/sqlparser/rewriter.go
@@ -4,17 +4,7 @@ package sqlparser
 
 //go:generate go run ./visitorgen/main -input=ast.go -output=rewriter.go
 
-import (
-	"reflect"
-)
-
-type replacerFunc func(newNode, parent SQLNode)
-
-// application carries all the shared data so we can pass it around cheaply.
-type application struct {
-	pre, post ApplyFunc
-	cursor    Cursor
-}
+import "fmt"
 
 func replaceAddColumnsAfter(newNode, parent SQLNode) {
 	parent.(*AddColumns).After = newNode.(*ColName)
@@ -1705,7 +1695,7 @@ func (a *application) apply(parent, node SQLNode, replacer replacerFunc) {
 		a.apply(node, n.Right, replaceXorExprRight)
 
 	default:
-		panic("unknown ast type " + reflect.TypeOf(node).String())
+		panic(fmt.Sprintf("unknown ast type %T", node))
 	}
 
 	if a.post != nil && !a.post(&a.cursor) {
@@ -1713,11 +1703,4 @@ func (a *application) apply(parent, node SQLNode, replacer replacerFunc) {
 	}
 
 	a.cursor = saved
-}
-
-func isNilValue(i interface{}) bool {
-	valueOf := reflect.ValueOf(i)
-	kind := valueOf.Kind()
-	isNullable := kind == reflect.Ptr || kind == reflect.Array || kind == reflect.Slice
-	return isNullable && valueOf.IsNil()
 }
